refactor(anti-spam): split react captcha exports from server barrel
Keep the root anti-spam entrypoint server-safe and expose CaptchaGuard from a dedicated react sub-entrypoint.
