+++ src/lib/anti-spam/react/index.ts
@@ -1,0 +1,3 @@
+export { CaptchaGuard } from "./captcha-guard";
+export type { CaptchaGuardProps } from "./captcha-guard";
+export type { CaptchaGuardProvider } from "../types";
